CO-PRIME TRIPLETS
This Java program generates and prints all primitive Pythagorean triplets ((a, b, c)) such that (c) is less than or equal to a user-defined maximum value. It uses Euclid's formula to find triplets, where integers m and n are chosen to ensure that m > n, m and n are coprime, and have opposite parity (one even, one odd). For each valid pair (m, n), it calculates:
- (a = m^2 - n^2)
- (b = 2 times m times n)
- (c = m^2 + n^2)
If (c)  meets the maximum limit, the triplet ((a, b, c)) is printed.